#180 Identify current ColorizedAction and perform appropriate command (start, stop, continue)
